+++ test/common/common.hpp
@@ -16,7 +16,7 @@ namespace found {
 
 ///// Definition of custom ASSERT_* statements /////
 
-#define DEFAULT_TOLERANCE 1e-3
+#define DEFAULT_TOLERANCE DECIMAL(1e-3)
 
 constexpr auto vectorEqual = [](const Vec2 &a, const Vec2 &b) {
     return abs(a.x - b.x) < DEFAULT_TOLERANCE && abs(a.y - b.y) < DEFAULT_TOLERANCE;
